# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: Search and list youth accounts
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [admin]
      summary: View a youth account
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    patch:
      operationId: updateUserStatus
      tags: [admin]
      summary: Activate or deactivate a youth account
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/opportunities:
    get:
      operationId: listAdminOpportunities
      tags: [admin, opportunities]
      summary: List all opportunities, including drafts
      responses:
        "200":
          description: Opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
    post:
      operationId: createOpportunity
      tags: [admin, opportunities]
      summary: Create an opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "201":
          description: Created opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /admin/opportunities/{id}:
    patch:
      operationId: updateOpportunity
      tags: [admin, opportunities]
      summary: Edit or publish an opportunity
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityUpdate"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [admin, opportunities]
      summary: Delete an opportunity
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: Opportunity deleted }
  /admin/programs:
    get:
      operationId: listAdminPrograms
      tags: [admin, programs]
      summary: List all programs, including drafts
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [admin, programs]
      summary: Create a B4P program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /admin/programs/{id}:
    patch:
      operationId: updateProgram
      tags: [admin, programs]
      summary: Edit or publish a program
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [admin, programs]
      summary: Delete a program
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: Program deleted }
  /admin/applications:
    get:
      operationId: listAdminApplications
      tags: [admin, programs]
      summary: Search and filter all program applications
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
      responses:
        "200":
          description: Applications with applicant profile
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminApplication"
  /admin/applications/{id}:
    patch:
      operationId: updateApplication
      tags: [admin, programs]
      summary: Update an application status and optional admin note
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminApplication"
...
  /admin/events:
    get:
      operationId: listAdminEvents
      tags: [admin, events]
      summary: List all events, including drafts
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [admin, events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /admin/events/{id}:
    patch:
      operationId: updateEvent
      tags: [admin, events]
      summary: Edit or publish an event
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [admin, events]
      summary: Delete an event
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: Event deleted }
...
  /admin/event-registrations:
    get:
      operationId: listEventRegistrations
      tags: [admin, events]
      summary: View event registrations
      parameters:
        - in: query
          name: eventId
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Event registrations with attendee profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminEventRegistration"
  /admin/notifications:
    post:
      operationId: publishNotification
      tags: [admin, notifications]
      summary: Publish an announcement notification to all active youth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Published notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationReceipt"
...
  /admin/contact-messages/{id}:
    patch:
      operationId: updateContactMessage
      tags: [admin, contact]
      summary: Update the handling status of a contact message
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageUpdate"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get editable organization and contact information
      responses:
        "200":
          description: Organization settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Edit About B4P and contact information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationSettings"
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
...
  schemas:
...
    Opportunity:
      type: object
      required: [id, title, category, organization, contactInfo, coverImage, description, requirements, location, deadline, applicationUrl, datePosted, published]
      properties:
        id: { type: integer }
        title: { type: string }
        category: { $ref: "#/components/schemas/OpportunityCategory" }
        organization: { type: string }
        contactInfo: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"], format: uri }
        description: { type: string }
        requirements: { type: string }
        location: { type: string }
        deadline: { type: ["string", "null"], format: date }
        applicationUrl: { type: ["string", "null"], format: uri }
        datePosted: { type: string, format: date-time }
        published: { type: boolean }
    OpportunityInput:
      type: object
      required: [title, category, organization, description, requirements, location, published]
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        category: { $ref: "#/components/schemas/OpportunityCategory" }
        organization: { type: string, minLength: 1, maxLength: 200 }
        contactInfo: { type: ["string", "null"], maxLength: 500 }
        coverImage: { type: ["string", "null"], format: uri }
        description: { type: string, minLength: 1 }
        requirements: { type: string }
        location: { type: string, minLength: 1, maxLength: 200 }
        deadline: { type: ["string", "null"], format: date }
        applicationUrl: { type: ["string", "null"], format: uri }
        published: { type: boolean }
...
    ProgramUpdate:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 200 }
        coverImage: { type: ["string", "null"], format: uri }
        description: { type: string, minLength: 1 }
        objectives: { type: string }
        eligibility: { type: string }
        location: { type: string, minLength: 1, maxLength: 200 }
        startDate: { type: ["string", "null"], format: date }
        endDate: { type: ["string", "null"], format: date }
        applicationDeadline: { type: ["string", "null"], format: date }
        published: { type: boolean }
...
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/UserProfile"
        - type: object
          required: [applicationCount, eventRegistrationCount]
          properties:
            applicationCount: { type: integer }
            eventRegistrationCount: { type: integer }
    UserStatusUpdate:
      type: object
      required: [active]
      properties:
        active: { type: boolean }